+++ cmd/ateapi/internal/store/atepg/schema.go
@@ -81,52 +81,52 @@ CREATE TABLE IF NOT EXISTS workers (
     proto    bytea NOT NULL
 );
 
--- Transactional change feed backing WatchWorkers. Events are appended in
+-- Transactional outbox backing WatchWorkers. Events are appended in
 -- the same transaction as the worker write and delivered by polling past
 -- an xid cursor;
 -- payload is one event-type byte followed by the binary Worker proto.
 --
 -- xid is the whole ordering: writeAndAppendChange appends exactly ONE row
--- per transaction (the only insert site), so every feed row has a distinct
+-- per transaction (the only insert site), so every outbox row has a distinct
 -- xid and a poll batch can never split a same-xid group. That invariant is
 -- load-bearing for the watch cursor and pinned by a test.
 --
--- Partitioned by created_at range (width: changeFeedPartitionInterval,
+-- Partitioned by created_at range (width: outboxPartitionInterval,
 -- kept <= retention so rows outlive it by at most one interval) so
 -- retention is a partition DROP — a metadata operation with no row
 -- deletes, dead tuples, or vacuum debt — instead of bulk DELETEs whose I/O
 -- competes with foreground traffic. The maintenance loop
--- (changeFeedMaintenance) creates upcoming partitions and drops expired
+-- (outboxMaintenance) creates upcoming partitions and drops expired
 -- ones; the DEFAULT partition only receives writes if partition creation
 -- ever stalls, and is truncated wholesale once maintenance notices
 -- (watchers that lose events to that resync via the trim mark).
 --
--- Partitions are UNLOGGED: the feed is ephemeral by design (cursors are
+-- Partitions are UNLOGGED: the outbox is ephemeral by design (cursors are
 -- not durable, subscriptions start "from now", and every consumer rebuilds
 -- from the workers table on resync), so paying WAL on every event — inside
 -- every worker-write transaction — buys nothing. Crash/failover truncates
 -- unlogged tables; see WatchWorkers for how watchers recover.
--- worker_changes_trim stays logged — the trim mark must survive a crash.
-CREATE TABLE IF NOT EXISTS worker_changes (
+-- worker_outbox_trim stays logged — the trim mark must survive a crash.
+CREATE TABLE IF NOT EXISTS worker_outbox (
     xid         xid8 NOT NULL DEFAULT pg_current_xact_id(),
     -- clock_timestamp(), not now(): now() is transaction-START time, so a
     -- slow transaction would route its event by a stale timestamp — worst
     -- case into an already-dropped partition (the DEFAULT partition would
-    -- catch it). The feed insert is the last statement before commit, so
+    -- catch it). The outbox insert is the last statement before commit, so
     -- statement time routes into the partition closest to commit time.
     created_at  timestamptz NOT NULL DEFAULT clock_timestamp(),
     payload     bytea NOT NULL
 ) PARTITION BY RANGE (created_at);
 
-CREATE INDEX IF NOT EXISTS worker_changes_xid ON worker_changes (xid);
+CREATE INDEX IF NOT EXISTS worker_outbox_xid ON worker_outbox (xid);
 
-CREATE UNLOGGED TABLE IF NOT EXISTS worker_changes_default PARTITION OF worker_changes DEFAULT WITH (autovacuum_enabled = off);
+CREATE UNLOGGED TABLE IF NOT EXISTS worker_outbox_default PARTITION OF worker_outbox DEFAULT WITH (autovacuum_enabled = off);
 
 -- Single-row high-water mark of retention: the greatest xid ever discarded
--- from worker_changes (dropped with an expired partition, or truncated
+-- from worker_outbox (dropped with an expired partition, or truncated
 -- with the DEFAULT partition). Watchers compare it against their cursor to
 -- detect exactly that unconsumed rows were discarded out from under them.
-CREATE TABLE IF NOT EXISTS worker_changes_trim (
+CREATE TABLE IF NOT EXISTS worker_outbox_trim (
     id   boolean PRIMARY KEY DEFAULT true CHECK (id),
     xid  xid8 NOT NULL
 );
